alterações para funcionamento do webhook

+++ src/TechLanches/Adapter/Driver/TechLanches.Adapter.API/Endpoints/PagamentoEndpoints.cs
@@ -47,7 +47,7 @@ namespace TechLanches.Adapter.API.Endpoints
             return Results.Ok(pagamento.Adapt<PagamentoStatusResponseDTO>());
         }
 
-        private static async Task<IResult> BuscarPagamento(int id, TopicACL topic, [FromServices] IPagamentoACLService pagamentoQrCodeACLService, [FromServices] IPagamentoService pagamentoService)
+        private static async Task<IResult> BuscarPagamento(long id, TopicACL topic, [FromServices] IPagamentoACLService pagamentoQrCodeACLService, [FromServices] IPagamentoService pagamentoService)
         {
             if (topic == TopicACL.merchant_order)
             {
